requirements.txt
# Core Framework
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-CORS==4.0.0
Flask-JWT-Extended==4.6.0
Flask-Bcrypt==1.0.1
Flask-Limiter==3.5.0
Flask-Caching==2.1.0

# Database
psycopg2-binary==2.9.9
SQLAlchemy==2.0.23
alembic==1.13.1

# Redis for caching and session
redis==5.0.1

# Data Validation
marshmallow==3.20.1
email-validator==2.1.0

# Utilities
python-dotenv==1.0.0
Werkzeug==3.0.1

# Production Server
gunicorn==21.2.0
gevent==23.9.1

# Monitoring and Logging
prometheus-flask-exporter==0.23.0

# Payment Gateway (Stripe for demo)
stripe==7.8.0

# Date/Time
python-dateutil==2.8.2

# Testing
requests==2.31.0
